HZN-1345: Refactor module structure
